add feature flag for the new and old sidebar
